requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
requests-html>=0.10.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
readability-lxml>=0.8.1
Pillow>=10.0.0
pytesseract>=0.3.10
openai>=1.0.0
google-generativeai>=0.3.0
tenacity>=8.2.0
pydantic>=2.0.0
validators>=0.22.0
python-dateutil>=2.8.2

# HuggingFace transformers for fake news detection model
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99
